pyproject.toml
[tool.poetry]
name = "sales-network"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^25.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
flake8 = "^7.1.2"
black = "^25.1.0"
isort = "^6.0.0"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.5.0"
drf-yasg = "^1.21.9"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
